Document-Based-GraphRag/Document-Based-GraphRag/requirements.txt
pymupdf>=1.20.0
python-dotenv
fastapi
uvicorn
requests
huggingface-hub==0.16.4
networkx>=3.1
graphviz>=0.20.1
neo4j==5.15.0
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl 
keybert==0.7.0
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.45.1
google-generativeai==0.3.2
